Implement query logic on BSSM
Summary:
While the previous diff added the "write logic" (of creating BSM), this one adds the "read logic" (reading data from BSM)

I added the `find_files_filter_basenames` method which searches for files with a given basename in the repo using the BSM.

To do that we need to cleverly (read `basename_suffix_skeleton_manifest.md` for more info) modify the query so that it becomes listing files in a subdirectory.

Notice that I added the function inside `RootBasenameSuffixSkeletonManifest`, and didn't implement a `pub fn id(&self)` like in all other manifests. This makes it hard for us to make mistakes by using the BSSM directly, which needs to be used carefully because of the modifications we make to the path.

This will be used on the next diffs.

Differential Revision: D38712937

fbshipit-source-id: b1fb4859338b76d0edc829474d93141956c9d69a