Add BasenameSuffixSkeletonManifest thrift and basic id types, no implementation
Summary:
This diff adds the basic thrift format for BasenameSuffixManifest (it contains a single field, just a sharded map similar to skeleton manifest, the secret is in the details).

It adds the code to parse it to/from rust, and the ids, including the custom sharded map node ids, which are possible due to D37112685 (https://github.com/facebookexperimental/eden/commit/62c2fd94cdf148a753d1038b5ea17550c1130b20).

It doesn't add any logic about creating or using a "basename suffix skeleton manifest", this will come in the next diffs.

Ideally this would be less boilerplate in the future. We have some helpful macros, but it could be way easier.

Reviewed By: mitrandir77

Differential Revision: D37113673

fbshipit-source-id: c8091e6440855e9c224a2096e915f27ca3704225