workingcopy: add workingcopy getter to repo
Summary:
We need to add a way to be able to get a workingcopy from a repo. Eventually we may be able to get many keyed by a path but for simplicity's sake let's stick to one now.

In the next few diffs I'm going to be looking at having the Eden fast path go through the workingcopy, starting with a new WorkingCopy CLI command. Therefore I'm happy to hardcode some things for the Eden case for now.

Reviewed By: DurhamG

Differential Revision: D38621169

fbshipit-source-id: eb6b16488afc1c705805ce3af7ce1c2013045054