cli_rs: collect redirections directories early in du
Summary:
By having a pattern where everything is collected upfront, and then some
computation is run, the code becomes slightly easier to follow as there is a
clear separation of concerns.

This also enables a later step where orphaned redirections will be collected
and printed/cleaned, having a set of all the known redirections will simplify
that step.

Reviewed By: kmancini

Differential Revision: D39005768

fbshipit-source-id: 5adb06c5ca1e69510a3ed8927bf47165215545e1