tests: crank up watchman min_acceptable_nice_value
Summary: In hg integration tests Watchman would fail to start if its nice value was above 0. Crank up its nice threshold to 999 so it never complains. I decided this was better than lowering watchman's nice value to 0. A test watchman is not so critical, and it seems appropriate to inherit the assigned nice value rather than overriding it.

Reviewed By: jordanwebster

Differential Revision: D38787274

fbshipit-source-id: 9c02da26fc0f3905a0c0570c9b20650a542cb7cb