simplify Makefile so `make deb` picks the right Ubuntu version
Summary:
Simplifies the build logic so the only thing `make deb` does is
run the `./packaging/debian/build_deb.sh` script. Note that the
role of `build_deb.sh` has expanded so that it:

- reads the Ubuntu version from `/etc/os-release`
- sets `PY_VERSION` and `GIT_DEB_DEP` accordingly
- runs `DESTDIR=install make install-oss`

This means that someone on either Ubuntu 20.04 or Ubuntu 22.04
can just run `VERSION=0.0.0 make deb` and it should do the right thing.

Reviewed By: quark-zju

Differential Revision: D39182615

fbshipit-source-id: 073d569a1bcecfb9fbeb78e8042ac18b0c1bb879