service: make sure globbing runs in the background when requested
Summary:
Now that the glob code relies exclusively on ImmediateFuture, the globs are
resolved eagerly during the Thrift call. This is usually what is desired,
except when a background glob is requested, in which case the intent is for
the Thrift call to complete as fast as possible and for the globbing to occur
in the background. By using the newly introduced makeNotReadyImmediateFuture
API, we can achieve this by keeping the rest of the code unchanged.

Reviewed By: genevievehelsel

Differential Revision: D37331421

fbshipit-source-id: 26685cbfbf5b229f0e987cda483b876a55aa998b