Implement repo-name encoding for sharded repos
Summary:
Currently, only mercurial repos are hosted by all Mononoke services and jobs. These repos are mostly alphanumeric (with sometimes `_` in their name). The git repos that are going to be mirrored in Mononoke have other special characters (e.g. `/`, `-`, `+`, `.`) in their names. ShardManager doesn't allow for such characters to be present in its configuration. This diff introduces a simple encoding scheme that allows for ShardManager-friendly repo names.

This diff depends on D39140863 which performs the actual encoding of repo-names.

Reviewed By: Croohand

Differential Revision: D39140281

fbshipit-source-id: 0e636ebd9e9112e974299fe5db63c1357feb2ac3