encode '/' and '+' in repo names in urls
Summary:
In order to support git repositories mirrored in mononoke without name mapping, we need to support slashes and plus signs in repo names.

This diff introduces a new crate that encodes and decodes repo names.

Usage of the crate is being introduced in toml config parsing, git_repoimport and edenapi client.

Reviewed By: mitrandir77

Differential Revision: D38975480

fbshipit-source-id: c04dd33f613bdec2605152872ef4b08b5e73143b