derived_data: Remove `BlobRepo` from utils
Summary:
The surface of this diff is a bit larger than what I would have liked, but the `DerivedUtils` trait is already polymorphic, so it was necessary for object safety that its functions take a type that's known at compile time: It is used in a context of dynamic polymorphism.

I tried to work around it with static polymorphism, but that didn't work as `derived_data_utils_impl` needs to return different types.

Instead, have the functions in `DerivedDataUtils` take the `RepoDerivedData` struct rather than an object implementing one of the `RepoDerivedDataRef` or `RepoDerivedDataArc` traits.

Change all the callers to pass the appropriate types.

Reviewed By: mitrandir77

Differential Revision: D39263429

fbshipit-source-id: 0c118484e835c4ccc5193d4de76c99b9e9dba1d4