add metadata middleware
Summary:
This middleware is a replacement for client_identity one. It will provide a ready building block for building the CoreContext for given app.  The ClientIdentity contents were just subset of Metadata anyway so why not use it directly.

NB: in the future we might want to also unify instantiating CoreContext but
this stack won't go that far.

Reviewed By: yancouto

Differential Revision: D38964580

fbshipit-source-id: 7f0b6ae57bc71fb4b6b591a94cfbf19fffac1e13