mononoke/bookmarks: log raw number of retries to ODS while writing to bookmarks_update_log
Summary:
While writing to `bookmarks`/`bookmarks_update_log`, we can have some retries of `RetryableError` before we hit success or any other error.

We already have a number of ODS stats here but calculating the raw number of when we retry the `RetryableError` is tedious: you need to either look at `{outcome}_attempt_count.avg.60` or `{outcome}_attempt_count.sum.60 - {event}.sum.60`  (and you need to do it for all possible `outcome`'s). Both of approaches don't work properly on heavily aggregated (old) ODS data and it is hard to tell which one is more accurate.

Since we're investigating if these retries are frequent and are indeed an issue in production, let's log their precise count instead of trying to derive it from other multiple metrics.

Reviewed By: yancouto

Differential Revision: D39280215

fbshipit-source-id: 15d19bc4901b9f7fa344f682fb43da0ceefd8765