clidispatch: add RequestContext convenience type
Summary:
This object contains commonly used cli facilities and will be provided to cli commands. At first it will serve to simplify messy fn signatures, but will continue to evolve further utility.

There is an inner CoreContext object that is not coupled to a particular CLI command. This can be passed around more willy nilly to other libraries.

Reviewed By: quark-zju

Differential Revision: D38882455

fbshipit-source-id: 73f1b163fbeb497ffc04060b4e40c703b3017b9a