cli_rs: allow `eden du` to measure orphaned redirections
Summary:
When a redirection is removed, or when a repository is recloned, some
redirections become orphaned. The user is for the most part unaware of this and
as a result, the orphaned redirection can consumme a large amount of disk
space, with no easy way to reclaim it.

Since redirections are created in the scratch directory, we can walk it and
deduce which directories were an EdenFS redirection. On macOS, redirections are
created differently and this diff doesn't attempt to handle them, this will be
tackled at a later time.

Reviewed By: mshroyer, kmancini

Differential Revision: D39076974

fbshipit-source-id: 6402a352e3a7054583794f6c42ad8720a982f3b4