hg: do not compute legacy proxy hash at construction
Summary:
Looking at our profiling, the prepareToStoreLegacy appears to be fairly hot due
to needing to compute sha1. Since the HgProxyHash discards the computed hash,
this is purely wasted CPU time and thus can be removed. According to the
profiling, 2/3 of the time spent constructing HgProxyHash is spent computing
sha1.

Reviewed By: chadaustin

Differential Revision: D39331915

fbshipit-source-id: ef4e3ec6172148d7bf3e945b830ad52c6e02e8ff