hg_sync_job: process batches sequentially
Summary:
The pipelining of batches of things to sync has led to a bug where we
accidentally cloned the `BookmarkOverlay` for each batch, resulting in the
values getting staler and staler as the process ran.

It's not immediately obvious how necessary this pipelining is: we build each
batch in parallel, and then push them (necessarily) one-by-one.  It is simpler
to do this sequentially.  We can re-introduce the pipelining if the performance
shows it is necessary.

Reviewed By: RajivTS

Differential Revision: D39278687

fbshipit-source-id: c743083ad76549ce2eed10d8600feb4a70c3ff43