BSSM: Change sentinel to fix ordering bug
Summary:
When we store a path in BSSM, we modify it. We used to add a `$` sentinel to it, but this diff changes it to use the basename.

So, `eden/mononoke/TARGETS` used to become `STEGRAT/eden/mononoke/$` and now becomes `STEGRAT/eden/mononoke/TARGETS`.

The reason is that this fixes ordering when listing files with the same basename.

Before, we would have `a/b/c` >  `a/b/a/c` , but `c/a/b/$`< `c/a/b/a/$`.

Changing the sentinel to something like `~`, on the other side of the ascii table, doesn't really fix it, only makes other cases problematic.

But by using the basename in the end, we keep relative order correct.

Reviewed By: markbt

Differential Revision: D39281796

fbshipit-source-id: bfbf83bcccd81ad65fc4234d37a1b43e13b0dfa5