crdump: respect "--unified 0" option
Summary:
The crdump extension treats the "unified" arg as an int (because it has an inty default value). If a user specified "--unified 0", the 0 was falsey and got changed to a default value later by some arg munging.

The "diff" command OTOH treats the "unified" param as a string (due to a default of ""), so the user specifying 0 becomes "0", which is truey and ends up sticking.

Fix by tweaking the arg munging to not treat an int 0 as empty. I don't think this should break things because there didn't seem to be any other int params handled in this code. The loosey-gooseynes of CLI flag types is problematic, but I don't feel like touching that in general right now.

Reviewed By: DurhamG

Differential Revision: D39406183

fbshipit-source-id: ecb62d654a0661fba5539a9e7f82dfe4c6eba59d