tests: repurpose test-fb-hgext-remotefilelog-linknodes.t for fastlog test
Summary:
linknodes are a tech-debt to be avoided and can be practically unusable slow.
Remove the test as we don't trigger linknodes paths in normal cases (i.e. no
copytracing and log uses fastlog).

In the near term I hope to make pathhistory the default to replace
linknodes+filenodes entirely.

The test is also the only test about fastlog. This diff removes the linknodes
part and repurpose the test for fastlog only.

Differential Revision: D38815652

fbshipit-source-id: a0ac9b536915c5ca2375461eb486ef3256deef3e