Treat tree as same as Blob for SetPathObjectId
Summary:
Why this would work?
Currently, what setPathObjectId doing for a blob is (for example, to set "a/b/c" to "blobDigest")
1. find parent treeInode, which is "a/b"
2. generate a fake tree with entry of "c"
3. call checkout of "a/b" inode.

In fact, 3 will ends up the following steps
1. insert the TreeEntry to "a/b" contents
2. update overlay.

However, for a tree (for example, to set "a/b/c" to "treeDigest")
1. create dir of "a/b/c"
2. call checkout of "a/b/c" inode with incoming tree of getTree result of "treeDigest"

2 will ends up the following steps.
1. get treeEntries from getTree result of "treeDigest"
2. add treeEntires into "a/b/c" inode contents list.
3. update "a/b/c"

However, if we assumes "a/b/c" will be insertion only, we can treat tree in the same way as blob as. In this case, we will
1. leave "a/b/c" unloaded until needed.
2. Save a call to CAS getTree
3. Save overlay updates.   Originally, creating "a/b/c" requires 6 overlay update, then another one for checkout on "a/b/c". Now, creating "a/b" requires 4 update and one for checkout on "a/b". This will save 2 overlay updates per SetPathObjectId call if it is a tree.

There is a downside doing this one: we need a special treatment for set on root tree, which equivalent to checkout the digest.

Reviewed By: chadaustin

Differential Revision: D37863687

fbshipit-source-id: 7941198d23ba13e465da623905cc681cba1fe8f5