add nodejs to the "Recommended" section of the .deb
Summary:
Add `Recommends: nodejs` to the `control` file for our `.deb` as Node is
not necessary for the core functionality of Sapling, but it is necessary for ISL.

Note that recommended packages are installed by default, so things should
still "just work" for ordinary users, though Ubuntu/Debian users have been
known to do:

```
sudo apt install --no-install-recommends nodejs
```

Though I assume they want to manage their own Node installation via `nvm`
or somesuch in those situations.

Note that now that we are installing the `nodejs` package, we need to update
`verify_release.sh` to set `DEBIAN_FRONTEND=noninteractive` so we don't
get stuck with timezone prompts when setting up the Docker image.

Reviewed By: jordanwebster

Differential Revision: D39493888

fbshipit-source-id: a8b62674273cb97dfeaf695fbbe5e1ff7e6a5501