clidispatch: change name for aliases
Summary: On Rust commands we were setting the list of aliases for Rust commands into some "name" variables and functions. For instance, we would set `status|st|sta|stat|statu` as the "name" for the Rust status command. This changes "name" for aliases in both variables and functions where a list of aliases might be accepted.

Reviewed By: quark-zju

Differential Revision: D39495554

fbshipit-source-id: 6dd843e3cc72fa8b0e42eacc15f180d6c7e62f6f