clidispatch: make dispatcher better handle commands.force-rust
Summary: Previously Rust commands had to set the command name when indicating they had to fall back to Python. For instance, if one wanted to prevent the Rust clone command falling back to Python when setting `commands.force-rust=clone`, one had to return a `errors::FallbackToPython("clone")` error. This makes the dispatcher use the first alias in the list of aliases as the main name for commands instead of the string in the `FallbackToPython` error.

Reviewed By: quark-zju

Differential Revision: D39452275

fbshipit-source-id: d03dc6c420241d661dbedff96baaf9bf2ff54457