add support to gate Rust commands behind sitevar
Summary:
When Chef runs, we read a sitevar from CPE_EDENFSCTL. This sitevar contains a mapping of experimental_rust_cmd -> rollout%. Chef then uses this % to write a json to /etc/eden

At run time, we read this file, `/etc/eden/edenfsctl_rollout.json`, and decide whether the command we were given is part of the list. If it's part of the list and it resolves to true, we should run the Rust version of the command. If it's not part of the list, we should run the python version.

To make this work when we don't have the edenfsctl_rollout.json on disk, we add a "default json" which we can update when other commands start rollingout.

The benefits of using a sitevar:

1) we can't use our .edenrc because we are too early in the command's invocation. We also don't want to do the unnecessary work of creating an EdenFsInstance{} in rust if we're going to fall back to Python anyway.
2) We can easily update the sitevar if anything goes wrong.
3) We can even add commands to the list that weren't previously considered "experimental". This is a generic solution for any oxidized command rollout.

Reviewed By: fanzeyi

Differential Revision: D38602957

fbshipit-source-id: f1142d3d8591759c4186bed7117389e15478b770