implement Rust `eden redirect add` helpers
Summary:
Add implementation for the Rust version of `eden redirect add` helpers in Rust.

The existing `eden redir add` implementation implements the is_buckd_running_for_path() logic incorrectly, so we will just copy that for now to keep it consistent. In reality, it should be looking for the buckd/pid file in a subdirectory of the repo root (not the repo root itself).

Reviewed By: fanzeyi

Differential Revision: D38231112

fbshipit-source-id: 164bb76922e4d228682d4b07a9cf04341a5f795d