doctor: add telemetry for problem descriptions
Summary:
The telemetry only contains the type of problem, and while this gives us a good
understanding of the kind of issue a user is facing, there isn't enough data to
be able to dig onto the issue itself without a user copy/pasting the output of
`eden doctor`. To improve on this, we can collect the error that is being
printed to the user and send it as a telemetry data which will give us more insight.

In particular, I'm hoping to better understand some of the errors seen on
Windows where EdenFS and the filesystem disagree about the state of
files/directories.

Reviewed By: genevievehelsel

Differential Revision: D39325412

fbshipit-source-id: 9b307b13f80fdaa3b972f69d2e5dac44ea935769