doctor: catch several exceptions when verifying inodes
Summary:
According to the telemetry, one of the highest reported issue is `eden doctor`
failing to check the mount for an unknown reason. These occur when an uncaught
exception is raised in one of the checker. When checking for loaded and
materialized inodes, there seems to be a couple of places where an exception
can occur without being caught, let's catch them and add a proper error to the
tracker. Once this is rolled out, we'll be able to better understand which of
these is failing so we can look into resolving the underlying issue.

Reviewed By: genevievehelsel

Differential Revision: D39296817

fbshipit-source-id: df6c06b7e81485dd0bb52875fb7e71605ea9819d