doctor: attempt to remediate MaterializedInodesHaveDifferentModeOnDisk
Summary:
According to our telemetry, this is the most seen issue on Windows. The root
cause for this is likely an issue with the dispatcher and while many issues
have been fixed there, not all of them have been found. Since self-remediating
can be done fairly easily (by just renaming the file/directory), let's do that
first. This should help engineers avoid having to reclone their repository
saving them a lot of time.

Reviewed By: genevievehelsel

Differential Revision: D39296816

fbshipit-source-id: 9cc16b5e9db48fbc3a99412145333b9087603911