Wait for replication lag in backfill_derived_data commands
Summary:
This was submitted on D39349941 (https://github.com/facebookexperimental/eden/commit/2ebfb3c0510afc59630abccd9141f873fa3cf9c5) and then revert due to errors in prod.

The fix is to use the repo_name we get in the function rather than infer it from matches, since that won't work in a shardmanager context.

Reviewed By: mitrandir77

Differential Revision: D39571926

fbshipit-source-id: 9b7ebe0f6848923595a65176c1126d17b68b6c40