status: show tracked-but-ignored files during status
Summary:
Creates a manifest matcher and uses it to unignore files that are
in the parent manifest but are match the ignore patterns.

This covers modified and removed files (i.e. files that have changes and are in
p1). It does not cover added files, which will be handled in the next diff.

Differential Revision: D39481483

fbshipit-source-id: e0aaf14c706f8c203bb2737321eb831b54b0a330