pathmatcher: shortcircuit exclude matcher during DifferenceMatcher
Summary:
In a DifferenceMatcher, if the include matcher says "match nothing",
then we can avoid even computing the exclude matcher result. The common use case
for the DifferenceMatcher is for ignore matchers. In the status case, the ignore
matcher can have some semi expensive logic that inspects the p1 manifest and the
treestate. This optimization will let us avoid looking at that data in any
non-ignore cases.

Differential Revision: D39482111

fbshipit-source-id: c1595413876c8f50c41dde49f217c2d2b18c3706