workingcopy: don't store a manifest on the Rust workingcopy
Summary:
Currently the Rust workingcopy struct stores a Manifest. We don't
really want to store a manifest, and instead we should resolve the appropriate
manifest based on the treestate parent. Let's start by storing a manifest
resolver object, which can be used to get a manifest when given a commit id.

In a later diff this will be used to refactor the workingcopy to support
multiple parents, to fix 'hg status' during merges.

Reviewed By: quark-zju

Differential Revision: D39520966

fbshipit-source-id: 30436ba3848cd0b5400f5a99e1f632d05e5edb41