model: use a constant time hash for ObjectId
Summary:
The ObjectCache is an extremely hot code path, dominating the glob query times.
When profiling it, hashing ObjectId shows up as a big component of the
ObjectCache runtime. By convention, we know that ObjectId contain a hash
already, and thus we can use that as the ObjectId hash directly.

Reviewed By: chadaustin

Differential Revision: D39399209

fbshipit-source-id: b8113102ac94445efed74dcdd88bb0b1f3c62591