fix Rust config dir selection
Summary:
Previously, we would simply check if --config-dir was supplied. If not, we would use the default config dir path.

This is incorrect, as we're supposed to check whether the cwd is located in an eden mount. If so, we should use the config dir for whatever Eden mount we're currently in.

Reviewed By: fanzeyi

Differential Revision: D39705519

fbshipit-source-id: 233520ff3046bfb4b004b919f917c0569c804e91