color: fix colors when using "less" pager
Summary:
When using a pager, we want to inherit the "is tty?" determination from before we fiddle with stdout to point to the pager. Python had some special logic for this we lost when we switched to Rust to answer "should color?". Streampager still worked because Rust handles that pager. "Less" is handled in Python, so Rust only saw that stdout wasn't a tty, so it disabled colors.

Interestingly, the "should color?" determination happens before we enable the pager. However, we re-initialize the "color" extension after starting a non-streampager pager, so that disables colors in this case. Fortunately, we can just remove the re-initialization because the motivating config "color.pagermode" is apparently no longer supported. This works around the issue because now we only ask Rust "should color?" before the pager is activated.

Reviewed By: sggutier

Differential Revision: D39749806

fbshipit-source-id: a5ce205200b555fbfcbadcdf030d5f2b9ab08b38