Fix a bug affecting ranges
Summary:
The drawdag crate allows specifying ranges of nodes in the ASCII representation of the graph, for example "A..E" will expand to "A-B-C-D-E". It does this by starting with the string "A" and repeatedly calling the function str_succ on it as long as it is lexicographically smaller than "E".

The problem with this is that calling str_succ does not always return a lexicographically larger string, which sometimes results in infinite loops and in other cases incorrect results like "A..AA" expanding to "A-B-AA". This diff fixes this by defining a function is_successor that takes two strings and determines if the first can ever be equal to the second if we repeatedly apply str_succ to it, and then replacing the lexicograpical comparisons with calls to this function.

Reviewed By: quark-zju

Differential Revision: D39690410

fbshipit-source-id: 0790befef1d7c88fbc34e84e593fe659dd19bab7