Add pushrebase logic and fix integration test issues
Summary: This diff adds the pushrebase logic to the land service. Now, the land service is fully working remotely. We also did some refactoring from previous diffs to better modularise the code. It means, we moved all code related to errors to the error.rs file and created functions for code that was too long to do conversions, for instance.

Reviewed By: yancouto

Differential Revision: D39748263

fbshipit-source-id: 6b9ad9681a0c01c21fad20587856d17b18ff5781