treestate: switch TreeState storage to Arc<Mutex<TreeState>>
Summary:
The current Rust TreeState logic exists in a bit of limbo. During a
Python command, the actual TreeState is stored in pytreestate bindings as an
Arc<Mutex<Option<TreeState>>>, but in Rust commands TreeState is stored as a
Rc<RefCell<TreeState>>. When Python calls a Rust function, it temporarily
extracts the TreeState from the Mutex<Option<>> and hands ownership to Rust,
which wraps it in Rc<RefCell<>>.

In an upcoming series of diffs I want to move the source of truth of the
TreeState to be in Rust, even for Python commands. To do so, let's unify the
ownership model into Arc<Mutex<>> for now. Later, once Python is completely
ignorant of TreeState storage, we can downgrade to an Rc<RefCell<>> or
something.

Reviewed By: jordanwebster

Differential Revision: D39660331

fbshipit-source-id: 6ca1ea730d48ce9f4bebef29094d2ea36fbc1928