treestate: move treestate file name decision into the treestate
Summary:
The treestate file name was currently being decided in multiple places,
like dirstate.py and the Rust clone crate. Let's make a new TreeState::new()
function that creates a new TreeState with a new name.

Let's also remove the filename tracking from dirstate.py, and just read it out
of the treestate. This will make a future diff possible, where we move the
source of truth for the treestate into the Rust working copy.

Reviewed By: jordanwebster

Differential Revision: D39714712

fbshipit-source-id: 2c91f61bd6b6b3527945b3e6ce4a233cf65485c2