edenapi: in tests, don't fail if edenapi.url is not present
Summary:
In production edenapi.url is always present, but some tests don't use
the modern client/server configuration. We want to start using the scmstore in
more places, as part of our move to Rust, so we need to be able to instantiate
a scmstore in non-edenapi test environments.

Let's allow this for now. Later we can fix the tests that don't work with this.

Reviewed By: jordanwebster

Differential Revision: D39714709

fbshipit-source-id: 422e26117c05f3374fad7c9e5d5e25a20cd32eb1