mononoke: blobrepo_hg: Rename changeset_exists to hg_changeset_exists
Summary:
In the next few diffs I am going to add `hg` into the names of
`BlobRepoHg` methods. This is to make it clear that the methods operate on hg
data and also to prevent name clashes between `BookmarksMaybeStaleExt` and
`BlobRepoHg` methods.

Rename `changeset_exists` to `hg_changeset_exists` and update callers.

Reviewed By: yancouto

Differential Revision: D39509517

fbshipit-source-id: 8d2d9ab7b960bd42cd79af59afe0d018623c668c