fix naming collision between function and file
Summary:
Because `github_repo` was added as a `templatekeyword`
function in `eden/scm/edenscm/ext/github/__init__.py`,
that caused problems when trying to import the file
named `github_repo.py` in the same folder.

Do the simplest thing and rename the file to `github_repo_util.py`
and update the `import` statements.

Reviewed By: DurhamG

Differential Revision: D39759919

fbshipit-source-id: 139f21e789296e1f2187b93017540b8daf94421c