pyrepo: store a pyworkingcopy on a pyrepo
Summary:
In this series of diffs we want to move the source of truth for the
working copy down into Rust. In a Python command, we'll need some place to store
it though, so let's put it on the py repo object.

In Python, there's largely assumed to be a 1:1 relationship between a repo and
the working copy, so storing them together is fine. In Rust, there may be a
1:many relationship between the repo and working copies, so we don't want to
store these together on the Rust repo object.

Differential Revision: D39715994

fbshipit-source-id: d7331677e3d6411f154272da0c44b5402eb55f28