Print estimated duration on tailer and backfill-all (per slice)
Summary:
`backfill` subcommand prints an estimate of how long it will take to finish backfilling, which is really cool.

We can't really do it on `backfill-all`, as we don't load all commits, so don't know how many are left. But we can at least do so for the same "slice". Which already provides at least some small help.

Reviewed By: mitrandir77

Differential Revision: D39815861

fbshipit-source-id: fe99941395574509b8aae88920445f2cd9004760