Allow changing replication lag config mid-wait
Summary:
Last time when we had an issue, we changed the config for waiting for replication from 15s to infinity. However, all the processes still continued waiting for it to go lower than 15s before finally changing the config.

This makes the config even more hot reloadable, as we can change the config mid-wait, and should prevent problems like that.

Notice this is not perfect. It will still wait for `poll_interval` until completion even if the config changes. But that's fine because we never set a high value for `poll_interval`.

Reviewed By: mitrandir77

Differential Revision: D39770073

fbshipit-source-id: 0f3fd0e560f0c9227921a20952af9f66fd7cf80d