Add wait_for_replication on backfill subcommand
Summary:
Currently, in the `backfill_derived_data` command, the `backfill-all` subcommand waits for replication lag to go down before continuing, but the `backfill` command does not. They do the same thing in different ways (what changes is the way to fetch the commits that need deriving), so both should back off the same way.

This diff makes it so the `backfill` subcommand also waits for replication lag to die down between derivation batches.

Reviewed By: mitrandir77

Differential Revision: D39815224

fbshipit-source-id: 3f2a35f746352cfce52d8d80a21e4a25f4ea4705