Add SCS endpoint to derive data for a batch of commits
Summary:
For remote gitimport, we need to be able to derive the `fsnode` data on a batch of commits at once to avoid unnecessary rederivation of data when processing repos with pathological graph structures.

This will do away with a performance pessimization that has hindered us from using remote gitimport in production so far.

The first step towards this goal is to expose the functionality of deriving data on a batch of commits from the SCS server.

* Add a `repo_prepare_commits` endpoint to SCS server, deriving the necessary data
* Expose this client through the SCS client's CLI API for ease of testing
* Test that the endpoint fails gracefully when it shoulds and succeeds when it should

Reviewed By: mitrandir77

Differential Revision: D39810812

fbshipit-source-id: a37d6028f248b1a3becacf907b9db308bde48db8