status: use WorkingCopy status instead of static status
Summary:
Now that we share a working copy between Rust and Python, we can just
call status on it, instead of building a working copy in the static status
invocation.

Also moved last_write to become a parameter to status. Since we no longer create
a WorkingCopy explicitly for the status, we don't have the opportunity to set it
at that time. It probably didn't make sense to set it at working copy creation time
anyway.

Reviewed By: jordanwebster

Differential Revision: D39753347

fbshipit-source-id: c03777c084d3eea5b9d6f0b0cb59126c19f06588