store: do in-place refreshing of Rust dag_commits
Summary:
The existing dag invalidation throws away the old
Arc<RwLock<DagCommits>> and replaces it with a new one. In a future diff we'll
be holding a copy of the Arc in the WorkingCopy/TreeState layer, so if we
invalidate the dag in the repo, we also need to do it in the WorkingCopy.

To do so, instead of throwing away the whole RwLock, let's just replace the
DagCommits inside the RwLock. This way, all Arc's get the new value.

Reviewed By: jordanwebster

Differential Revision: D39777413

fbshipit-source-id: 0e2b3ec172c4c0ae71003cae2a05a63306f6fa67