store: allow refreshing Rust file and tree stores
Summary:
In a future diff we'll enable using the Rust status logic for Python
commands. This requires that the Rust commit/file/tree stores remain in sync
with the Python ones, even as the Python ones are mutated. Ideally we'd just use
the same store instances for Python and for Rust, but that's a bit more work and
somewhat risky. For now, let's just add invalidation methods on the stores.

Note, even in the future we may need invalidation methods, since it's possible
for an external process to change the repo and the internal repo instance will
need to pick up those changes.

Reviewed By: jordanwebster

Differential Revision: D39777412

fbshipit-source-id: 21ba25444a67142c6757b3446305864d2b6a520d