fix mercurial tests if executed with buck2
Summary:
fix mercurial tests if executed with buck2

Many of mercurial unit tests are failing with buck2.

As I discovered, this code path that is used for test compatibility with buck
doesn't work for buck2 because some naming convention has been changed.

With buck, the main hg binary is a _bash_ script setting up a bunch of things
instead of a standalone single executable. It is named with .sh for buck1 and
just hg for buck2.

Reviewed By: krallin

Differential Revision: D39923275

fbshipit-source-id: 8c5281150943e288b6ea78e77e9edc8ee151a051