Support Latin1 encoding in gitimport
Summary:
Before this change, we would fail to import a "ISO-8859-1" encoded commit, which would prevent importing some git repos, such as ariane in aosp.

For now, explicitly support Latin1 as it is easy to perform the conversion by hand and unblocking the current usecase is the first priority.

In a future diff, we will use a crate such as `encoding_rs` to support all encodings.

Reviewed By: mitrandir77

Differential Revision: D39931400

fbshipit-source-id: c3f68f529c1aa5bb7b6468a81641b06f7fcd2928