make clone respect starting eden
Summary:
eden clone trys to start eden if it isn't healthy. starting doesn't count as
healthy. lets wait for eden to start instead of starting it in clone.

This should get rid of more flock errors on windows and some sort of pid file
error on linux and mac as well. Xavier just noticed this flock error today.

There might be more of these healthy/starting check issues in the cli. will
audit for them.

Reviewed By: xavierd

Differential Revision: D39765586

fbshipit-source-id: 8949ccb7d4bbba8ee8df607d37c767b055113398