eagerepo: check trees and files references
Summary:
EagerRepo should not be lazy. That was previously not enforced.

This diff enforces the check on changing commits (add commit or update
reference) to ensure that the EagerRepo is not lazy.

The check on commit insertion can be made smarter (O(working copy) ->
O(changes)) performance-wise. That requires API changes that are out
of scope in this diff.

Differential Revision: D39525541

fbshipit-source-id: 1955bffc69b8cecea93189d48794ef327ddc3ef4