derived_data: rename backfill_batch to derive_exactly_batch
Summary:
Looking at [the `DerivedDataManager` interface](https://www.internalfb.com/intern/rustdoc/eden/mononoke/derived_data:derived_data_manager/derived_data_manager/manager/struct.DerivedDataManager.html), it is not very clear what "backfilling" means in this context.

The types of operations that are available are:
* deriving data
* fetching derived data
and these operations can be done on a single changeset or on a batch of changesets.

The reason the method was originally named `backfill_batch` is because of its use in `cmds/backfill_derived_data`.
This couples the interface to one of its specific users.

From the original documentation,
```
The difference between “backfill” and “derive”, is that for backfilling, the caller must have arranged for the dependencies and ancestors of the batch to have already been derived. If any dependency or ancestor is not already derived, an error will be returned.
```

The name `derive_exactly` reflects that difference more accurately.

Note: there is another discrepancy in the API:
For the case of a single changeset,
* `derive` derives the data (and any required dependency data) and returns it.
* `fetch` returns the derived data if it exists, or errs.

For a batch of changesets,
* `derive_exactly`  derives the data, but only return statistics on the derivation.
* `fetch` returns the derived data if it exists or errs.

That difference is due to the batch nature of the operation: it makes sense
to return statistics on a batch of changesets, but it doesn't make sense for a single commit.

Also remove the `_dangerous` suffix in `DerivedUtils::derive_exactly_batch` since "dangerous" used to mean we didn't check the pre-conditions, but the only implementer: `DerivedUtilsFromManager` does perform the check now.

Reviewed By: mitrandir77

Differential Revision: D39883708

fbshipit-source-id: f40cf7e21a38f06ebbad0743b5dda022b555625c