Fix incorrect EOF value in TreeInode::readdirImpl
Summary:
If we run out of NFS buffer space while trying to add the final directory entry
in TreeInode::readdirImpl we would incorrectly return true, indicating that
we've returned the entire listing.

This means that on NFS we would incorrectly send an EOF to the client, which
would then not send a subsequent, offset readdir request.  Because of this, for
certain NFS-mounted Eden directory contents, calling `ls` would result in a
single file being missing from the listing.

Reviewed By: chadaustin

Differential Revision: D39911610

fbshipit-source-id: 96a8fc7958a8e800ff7f2489c968f2dde23a86da